New optional GUI
Made many changes to support a QT6 GUI.
A GUI is now optionally provided and can be used instead of the CLI.
These changes incude:
A raise command flag to trigger errors in the gui.
Update of the builder to skip tkinter.
Removal of the univesal pyz build as it wont work without the code files.
The added gui also rememembes settings.
A icon for the GUI is also in the repo.
Updated setup.py to include qt6 for the gui.